moved transformSquareIntoArrayNotation into new file
